Handle invalid xml
It was failing on a <br> in raw_html because it's not closed with a
slash. Broke up the fb html body fragment from the article body which is
were the bad tag is, but didn't cause an issue.  It was only when the
second fragment ran.  So not seeing issues with other feeds.