Support single ANS types like story  AIO-847
The previous ANS change only fixed navigation, not story.  THat fix only looked if
content_elements existed.  But both results set and story contain content_elements
so changed it to look at type.  If results or collection use content_elements, if children
use it, otherwise wrap in an array and convert the top level items