Support xmlbuilder2 3.x  AIO-677
xmlbuilder2@3.0.0 fixed the mixed content bug, but 3.x changes the
way some tags are encoded.  Tags like <ThE_bOdY_gOeS_hErE/> are added
to the content then replaced with the customFields, this prevents the
tags from becoming encoded like &lt;script&gt;
But 3.x is encoding those so I had to switch them to comments <!-- -->
which it doesn't mess with.