Sitemaps at root  AIO-767
To support redirects for sitemaps at root we need allow them to use dashes instead of slashes and
add .xml to end of url with feedExtension
feedExtension - optional if set appends the value to the end of the url otherwise uses '/'
sitemapAtRoot - converts slashes to dashs for sitemap-section-index
In feeds-source-content-api-block
Sitemap-at-root - converts dashes back to slashes.  If they have sections with dashes in them,
those must be setup with their own resolvers.